Move ImportBatch() to the TargetYugabyteDB.
